requirements.txt
requests
beautifulsoup4
urllib3==1.26.16
selenium==4.17.2
python-telegram-bot==20.6
python-dotenv==1.0.1



